refactor: organize jarvis package
