+++ jarvis/__init__.py
@@ -1,9 +1,17 @@
 """Jarvis calendar assistant package."""
 
 from .services.calendar_service import CalendarService
-from .logger import JarvisLogger
-from .config import JarvisConfig
-from .log_viewer import LogViewerGUI
+from .logging import JarvisLogger, LogViewerGUI
+from .core import (
+    JarvisConfig,
+    JarvisSystem,
+    create_collaborative_jarvis,
+    DEFAULT_PORT,
+    LOG_DB_PATH,
+    ExecutionResult,
+    BaseRegistry,
+    FunctionRegistry,
+)
 from .ai_clients import (
     AIClientFactory,
     BaseAIClient,
@@ -12,28 +20,17 @@ from .ai_clients import (
 )
 from .agents.agent_network import AgentNetwork
 from .agents.calendar_agent import CollaborativeCalendarAgent
-
-try:  # Optional import to avoid heavy dependencies during module import
-    from .main_jarvis import JarvisSystem, create_collaborative_jarvis
-except Exception:  # pragma: no cover - optional dependency may be missing
-    JarvisSystem = None
-    create_collaborative_jarvis = None
 from .protocols import Protocol, ProtocolStep
 from .protocols.registry import ProtocolRegistry
-from .registry import BaseRegistry, FunctionRegistry
 from .protocols.executor import ProtocolExecutor
 from .protocols.builder import create_from_file
-from .constants import (
-    DEFAULT_PORT,
-    LOG_DB_PATH,
-    ExecutionResult,
-)
-from .performance import PerfTracker, track_async
+from .utils.performance import PerfTracker, track_async
 
 PicovoiceWakeWordListener = None
 OpenAITTSEngine = None
 ElevenLabsTTSEngine = None
 VoiceInputSystem = None
+
 __all__ = [
     "CalendarService",
     "AIClientFactory",
