+++ jarvis/agents/orchestrator_agent/__init__.py
@@ -11,9 +11,9 @@ from ..base import NetworkAgent
 from ..message import Message
 from ..task import Task
 from ...ai_clients import BaseAIClient
-from ...logger import JarvisLogger
+from ...logging import JarvisLogger
 from ...utils import extract_json_from_text
-from ...performance import track_async
+from ...utils.performance import track_async
 
 
 class OrchestratorAgent(NetworkAgent):
