+++ jarvis/core/__init__.py
@@ -1,0 +1,22 @@
+from .config import JarvisConfig, UserConfig, FeatureFlags
+from .builder import JarvisBuilder, BuilderOptions
+from .system import JarvisSystem, create_collaborative_jarvis
+from .constants import DEFAULT_PORT, LOG_DB_PATH, ExecutionResult
+from .profile import AgentProfile
+from .registry import BaseRegistry, FunctionRegistry
+
+__all__ = [
+    "JarvisConfig",
+    "UserConfig",
+    "FeatureFlags",
+    "JarvisBuilder",
+    "BuilderOptions",
+    "JarvisSystem",
+    "create_collaborative_jarvis",
+    "DEFAULT_PORT",
+    "LOG_DB_PATH",
+    "ExecutionResult",
+    "AgentProfile",
+    "BaseRegistry",
+    "FunctionRegistry",
+]
