+++ jarvis/core/registry.py
