+++ jarvis/io/output/tts/elevenlabs.py
@@ -7,9 +7,9 @@ from typing import Optional
 import httpx
 
 from ...utils.audio import play_audio_bytes
-from ....logger import JarvisLogger
+from ....logging import JarvisLogger
 from .base import TextToSpeechEngine
-from ....performance import get_tracker
+from ....utils.performance import get_tracker
 
 
 class ElevenLabsTTSEngine(TextToSpeechEngine):
