--- jarvis/logger.py
@@ -1,3 +1,0 @@
-from .loggers.jarvis_logger import JarvisLogger
-
-__all__ = ["JarvisLogger"]
