--- jarvis/loggers/__init__.py
@@ -1,3 +1,0 @@
-from .jarvis_logger import JarvisLogger
-
-__all__ = ["JarvisLogger"]
