+++ jarvis/logging/__init__.py
@@ -1,0 +1,4 @@
+from .jarvis_logger import JarvisLogger
+from .log_viewer import LogViewerGUI
+
+__all__ = ["JarvisLogger", "LogViewerGUI"]
