+++ server/main.py
@@ -8,7 +8,7 @@ from dotenv import load_dotenv
 
 
 from jarvis import JarvisLogger, JarvisSystem, JarvisConfig
-from jarvis.constants import DEFAULT_PORT
+from jarvis.core import DEFAULT_PORT
 from server.database import init_database, close_database
 from server.routers.jarvis import router as jarvis_router
 from server.routers.auth import router as auth_router
